add tracestate support
Add opentelemetry_trace_state_* function to create tracestate.

Add opentelemetry_span_trace_state_get() to read span tracestate.

Add parent_root sampler. The sampler is used to pass custom
tracestate to the root span. When the span is not a root then
the tracestate is copied from the parent span.